refactor(server)!: continue server v1 removal
