+++ x/genutil/client/cli/util.go
@@ -1,4 +1,4 @@
-package server
+package cli
 
 import (
 	"errors"
